--- backend/fittoring/src/main/java/fittoring/mentoring/CategoryMentoringRepository.java
@@ -1,6 +1,0 @@
-package fittoring.mentoring;
-
-import org.springframework.data.repository.ListCrudRepository;
-
-public interface CategoryMentoringRepository extends ListCrudRepository<CategoryMentoring, Long> {
-}
