--- backend/fittoring/src/main/java/fittoring/mentoring/ReservationRepository.java
@@ -1,6 +1,0 @@
-package fittoring.mentoring;
-
-import org.springframework.data.repository.ListCrudRepository;
-
-public interface ReservationRepository extends ListCrudRepository<Reservation, Long> {
-}
