+++ backend/fittoring/src/main/java/fittoring/mentoring/business/model/CategoryMentoring.java
@@ -1,6 +1,5 @@
-package fittoring.mentoring;
+package fittoring.mentoring.business.model;
 
-import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -12,18 +11,19 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "offering")
+@Table(name = "category_mentoring")
 @Entity
-public class Offering {
+public class CategoryMentoring {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false)
-    private String content;
-
     @ManyToOne
     @JoinColumn(nullable = false)
     private Mentoring mentoring;
+
+    @ManyToOne
+    @JoinColumn(nullable = false)
+    private Category category;
 }
