+++ backend/fittoring/src/main/java/fittoring/mentoring/business/model/Certificate.java
@@ -1,23 +1,29 @@
-package fittoring.mentoring;
+package fittoring.mentoring.business.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "category")
+@Table(name = "certificate")
 @Entity
-public class Category {
+public class Certificate {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, unique = true)
+    @Column(nullable = false)
     private String title;
+
+    @ManyToOne
+    @JoinColumn(nullable = false)
+    private Mentoring mentoring;
 }
