+++ backend/fittoring/src/main/java/fittoring/mentoring/business/model/Mentoring.java
@@ -1,29 +1,36 @@
-package fittoring.mentoring;
+package fittoring.mentoring.business.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "certificate")
+@Table(name = "mentoring")
 @Entity
-public class Certificate {
+public class Mentoring {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @Column(nullable = false)
-    private String title;
+    private String mentorName;
 
-    @ManyToOne
-    @JoinColumn(nullable = false)
-    private Mentoring mentoring;
+    @Column(nullable = false, unique = true)
+    private String mentorPhone;
+
+    @Column(nullable = false)
+    private int price;
+
+    private Integer career;
+
+    private String content;
+
+    @Column(nullable = false)
+    private String introduction;
 }
