+++ backend/fittoring/src/main/java/fittoring/mentoring/business/model/Offering.java
@@ -1,5 +1,6 @@
-package fittoring.mentoring;
+package fittoring.mentoring.business.model;
 
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -11,19 +12,18 @@ import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "category_mentoring")
+@Table(name = "offering")
 @Entity
-public class CategoryMentoring {
+public class Offering {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @ManyToOne
-    @JoinColumn(nullable = false)
-    private Mentoring mentoring;
+    @Column(nullable = false)
+    private String content;
 
     @ManyToOne
     @JoinColumn(nullable = false)
-    private Category category;
+    private Mentoring mentoring;
 }
