+++ backend/fittoring/src/main/java/fittoring/mentoring/business/model/Reservation.java
@@ -1,4 +1,4 @@
-package fittoring.mentoring;
+package fittoring.mentoring.business.model;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
