+++ backend/fittoring/src/main/java/fittoring/mentoring/business/repository/CategoryMentoringRepository.java
@@ -1,0 +1,9 @@
+package fittoring.mentoring.business.repository;
+
+import fittoring.mentoring.business.model.CategoryMentoring;
+import org.springframework.data.repository.ListCrudRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CategoryMentoringRepository extends ListCrudRepository<CategoryMentoring, Long> {
+}
