+++ backend/fittoring/src/main/java/fittoring/mentoring/business/repository/CertificateRepository.java
@@ -1,5 +1,6 @@
-package fittoring.mentoring;
+package fittoring.mentoring.business.repository;
 
+import fittoring.mentoring.business.model.Certificate;
 import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
