+++ backend/fittoring/src/main/java/fittoring/mentoring/business/repository/ImageRepository.java
@@ -1,5 +1,6 @@
-package fittoring.mentoring;
+package fittoring.mentoring.business.repository;
 
+import fittoring.mentoring.business.model.Image;
 import org.springframework.data.repository.ListCrudRepository;
 import org.springframework.stereotype.Repository;
